Give "now"-based tests a greater chance of passing
If these tests run too close to the turn of a second, then the different
calls to `Time.now` will return a different second and the test will
fail.

This changes the tests to wait until `Time.now` is more than 0.1 seconds
from the turn of the second. This should give us enough time to call
`Time.now` twice in the same second.