Add test coverate for "now"-based timestack tests
There are three ways that a TimestackItem can freeze time based on the
current time: with no arguments, with a single Integer or a single
Float. The tests for the single Integer or Float only test if the
argument is 0.

This commit adds tests for different values of the argument to
double-check that the time value is manipulated correctly, and adds a
test for when there are no arguments.